refactor Vault store
