+++ src/index.ts
@@ -1,3 +1,3 @@
 export { GCPPrivateKeyStore } from './lib/gcp/GCPPrivateKeyStore';
 
-export { VaultPrivateKeyStore } from './lib/vault/vaultPrivateKeyStore';
+export { VaultPrivateKeyStore } from './lib/vault/VaultPrivateKeyStore';
